Fixing calculation of points along pion's helix.

+++ include/Helix.hpp
@@ -58,7 +58,7 @@ public:
   inline int      GetNpionPoints(){return nPionPoints;}
   inline int      GetNregularPoints(){return nRegularPoints;}
   
-  inline int      GetNcycles(){return (/*sgn(slope) */ pixelBarrelZsize - origin->GetZ())/(slope*2*TMath::Pi());}
+  inline double   GetNcycles(){return (/*sgn(slope) */ pixelBarrelZsize - origin->GetZ())/(slope*2*TMath::Pi());}
   
   /// Calculates average of the squared distances between points (hits) and the helix
   double GetChi2();
